[FIX] website_sale_wishlist: fix pricelist assignation
Steps to reproduce:
- Install website_sale_wishlist,stock,sale_management,contacts
- General Settings>Sales>Multiple Sales Prices per Product
- Create or edit a product, assign 2 different prices in pricelists A and B
- Publish it to the website
- If your user is Mitchell Admin, go to its related res.partner, Sales &
Purchases tab/Pricelist, choose pricelist B
- Create a public user and assign it pricelist A
- Go to the website/shop
- Reach out the product and add to the wishlist

When the current user is internal the product page will display price of B,
but if you add the product to the wishlist, whishlist displays public price
Same for the portal user.
This happend because to avoid returning 500 when the product is
unpublished c93f371d67def785f5b6d90f96eed8343e962203 the product is
filtered using sudo() but the recordset is also returned with the sudo
environment. Using it just to filter solve the issue

opw-2166484

closes odoo/odoo#43617

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>